requirements.txt
asgiref==3.8.0
astroid==2.3.3
backports.zoneinfo==0.2.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==4.2.11
django-crispy-forms==2.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pillow==10.2.0
requests==2.22.0
six==1.13.0
sqlparse==0.4.4
sslcommerz-python==0.0.7
typed-ast==1.4.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==1.25.7
wrapt==1.11.2
